RP/requirements.txt
djangorestframework
alabaster==0.7.12
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
coverage==4.5.2
Django==2.1.7
docutils==0.14
future==0.17.1
idna==2.8
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mock==2.0.0
more-itertools==6.0.0
msgpack-python==0.5.6
packaging==19.0
pbr==5.1.2
pluggy==0.8.1
py==1.7.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.3.1
pytest==4.2.1
pytest-cov==2.6.1
pytz==2018.9
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
urllib3==1.24.1
